shell+libc: `gpio` command implemented, changed `pow` to use integers

+++ include/libc.h
@@ -7,4 +7,4 @@ void* memcpy(void* dest, const void* src, size_t n);
 size_t strlen(const char* str);
 int strcmp(const char* s1, const char* s2);
 int strncmp(const char* s1, const char* s2, size_t n);
-double pow(double base, double exponent);
+int pow(int base, int exponent);
